Moving test around to be more manageable going forward, making K2Tree fields private in prep for api stabilisation
